Rename and commit
